+++ lib/index.ts
@@ -15,24 +15,21 @@ export { defaultIcon };
 export function getIcon(filename: string): Icon {
 
   if (typeof filename === 'string') {
+    
+    // Use lower case filename for matching
+    const lcFilename = filename.toLowerCase();
 
     // Check complete filename first
-    let icon = iconDefs.icons.find(i => i.files && i.files.some(f => f.localeCompare(filename, undefined, { sensitivity: 'accent' }) === 0));
+    let icon = iconDefs.icons.find(i => i.files && i.files.some(f => f === lcFilename));
     if (icon) return icon;
 
-    // The extensions, can match multiple which leads to this quite inefficient way of matching.
-    const matches = iconDefs.icons.filter(i => i.extensions && i.extensions.some(ext => filename.endsWith('.' + ext)));
-    if (matches.length > 0) {
-      let matchLength = 0;
-      for (const match of matches) {
-        for (const ext of match.extensions) {
-          if (ext.length > matchLength && filename.endsWith('.' + ext)) {
-            matchLength = ext.length;
-            icon = match;
-          }
-        }
-      }
-    }
+    // Extensions can have multiple matches, select the longest matching extension.
+    let matchLength = 0;
+    iconDefs.icons.forEach(i => 
+      i.extensions?.forEach(ext => 
+        ext.length > matchLength && lcFilename.endsWith('.' + ext) && (icon = i) && (matchLength = ext.length)
+      )
+    );
     if (icon) return icon;
 
   }